+++ types/data-formats/output-data.d.ts
@@ -1,4 +1,5 @@
 import {BlockToolData} from '../tools';
+import {BlockTuneData} from '../block-tunes/block-tune-data';
 
 /**
  * Output of one Tool
@@ -8,6 +9,10 @@ import {BlockToolData} from '../tools';
  */
 export interface OutputBlockData<Type extends string = string, Data extends object = any> {
   /**
+   * Unique Id of the block
+   */
+  id?: string;
+  /**
    * Tool type
    */
   type: Type;
@@ -15,6 +20,11 @@ export interface OutputBlockData<Type extends string = string, Data extends obje
    * Saved Block data
    */
   data: BlockToolData<Data>;
+
+  /**
+   * Block Tunes data
+   */
+  tunes?: {[name: string]: BlockTuneData};
 }
 
 export interface OutputData {